fix: split currency fetch out of currency-utils to avoid browser client in server bundle
- Move fetchAndValidateCurrency (Supabase + logger imports) from
  currency-utils.ts into new currency-preference.ts; currency-utils
  is now a pure formatting/constants module safe to import in server
  components
- Init currencyLoading=true so consumers never see a false-loading
  window between auth resolving and the first currency fetch
- Add cancellation guard to refreshFetchedPrefs in useUserPreferences
  to prevent stale setState on unmount or fetchUserId change

Generated-By: PostHog Code
Task-Id: caa03bb9-705f-4aa7-9046-e28aa1482e92
